engine/requirements.txt
watchdog
google-generativeai
moviepy<2.0
python-dotenv
requests
